requirements.txt
crewai
langchain
langchain-groq
fastapi
uvicorn
pydantic
httpx
requests
groq
chromadb
flask
python-dotenv